requirements.txt
ddsmoothing-python>=1.0.0
torch>=1.6.0
torchvision>=0.7.0
numpy>=1.17.0
tqdm>=4.44.0
scipy>=1.5.0
statsmodels>=0.11.1